🤖 Fix signature positioning with accurate coordinate transformation
- Implement proper coordinate system conversion (Flutter pixels to PDF points)
- Add SignaturePosition class with PDF-to-viewer coordinate transformation
- Fix coordinate system origin differences (top-left vs bottom-left)
- Account for PDF aspect ratio and viewer scaling
- Update backend to use Double precision for coordinates
- Add coordinate debugging information dialog
- Remove debug print statements and clean up code
- Fix compilation errors in backend controller and service
- Improve signature positioning accuracy and reliability

+++ backend/src/main/java/com/firmador/backend/service/DigitalSignatureService.java
@@ -103,10 +103,10 @@ public class DigitalSignatureService {
         
         // Set signature rectangle
         Rectangle signatureRect = new Rectangle(
-            signatureRequest.getSignatureX(), 
-            signatureRequest.getSignatureY(), 
-            signatureRequest.getSignatureWidth(), 
-            signatureRequest.getSignatureHeight()
+            signatureRequest.getSignatureX().floatValue(), 
+            signatureRequest.getSignatureY().floatValue(), 
+            signatureRequest.getSignatureWidth().floatValue(), 
+            signatureRequest.getSignatureHeight().floatValue()
         );
         appearance.setPageRect(signatureRect).setPageNumber(signatureRequest.getSignaturePage());
